+++ examples/multi-language-i18next/src/views/pages/home/main.js
@@ -1,0 +1,1 @@
+console.log('\n>> Home page');